metastore/requirements/common.txt
Django==1.7.2
MySQL-python==1.2.5
django-admin-bootstrapped==2.3.5
django-tastypie==0.12.1
python-dateutil==2.4.1
python-mimeparse==0.1.4
six==1.9.0
wsgiref==0.1.2
django-jsonfield==0.9.13
scales==1.0.9
gunicorn==19.3
scribe-logger==1.1
django-zipkin==0.0.2
gevent==1.0.2
django-grappelli==2.6.5
python-ldap==2.4.13
django-ldap-sync==0.1.1
django-auth-ldap==1.2.7
django-historicalrecords==1.1
django-reversion==1.10.1
django-reversion-compare==0.6.1
requests==2.9.1
django-tastypie-swagger==0.1.4
GitPython==1.0.1
django-object-actions==0.7.0
enum==0.4.6
croniter==0.3.13
youtube-dl==2017.2.24.1

